[BUGFIX] Remove superfluous code fragments
