+++ Web/typo3conf/ext/sessions/Classes/Domain/Repository/AbstractSessionRepository.php
@@ -1,6 +1,8 @@
 <?php
 namespace TYPO3\Sessions\Domain\Repository;
 
+use TYPO3\CMS\Extbase\Persistence\QueryResultInterface;
+
 abstract class AbstractSessionRepository extends \TYPO3\CMS\Extbase\Persistence\Repository
 {
     /**
@@ -11,5 +13,27 @@ abstract class AbstractSessionRepository extends \TYPO3\CMS\Extbase\Persistence\
     {
         $query = $this->createQuery();
         return $query->matching($query->in('uid', $uids))->execute();
+    }
+
+    /**
+     * @return array()
+     */
+    public function findAllFlat() {
+
+        $query = $this->createQuery();
+
+        return $query->execute(true);
+    }
+
+    /**
+     * @return QueryResultInterface
+     */
+    public function getAcceptedSessionsOrderByVoteCount(){
+
+        $query = $this->createQuery();
+
+        $query->setOrderings(array('votes' => \TYPO3\CMS\Extbase\Persistence\QueryInterface::ORDER_ASCENDING));
+
+        return $query->execute();
     }
 }
